requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==4.2.14
django-filter==24.3
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.3
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.1
